fix: Entity PK 네이밍 규칙 통일, 계좌금액 타입 변경

+++ src/main/java/domain/account/entity/SaveAccount.java
@@ -31,7 +31,7 @@ public class SaveAccount extends BaseEntity {
 	private int saveAccountPassword;
 
 	@Column(name = "save_accounts_amount", nullable = false)
-	private int saveAccountAmount;
+	private long saveAccountAmount;
 
 	@Builder
 	public SaveAccount(String saveAccountNumber, int saveAccountPassword, int saveAccountAmount){
